+++ src/main/java/com/shy_polarbear/server/domain/auth/repository/rdb/RefreshTokenRepository.java
@@ -1,9 +1,11 @@
 package com.shy_polarbear.server.domain.auth.repository.rdb;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
+@Repository
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
 
     Optional<RefreshToken> findByRefreshToken(String refreshToken);
