+++ src/main/java/com/shy_polarbear/server/domain/auth/service/AuthService.java
@@ -3,13 +3,13 @@ package com.shy_polarbear.server.domain.auth.service;
 import com.shy_polarbear.server.domain.auth.exception.AuthException;
 import com.shy_polarbear.server.domain.auth.dto.response.LogoutResponse;
 import com.shy_polarbear.server.domain.auth.jwt.provider.ProviderType;
+import com.shy_polarbear.server.domain.auth.repository.rdb.RefreshToken;
+import com.shy_polarbear.server.domain.auth.repository.rdb.RefreshTokenRepository;
 import com.shy_polarbear.server.domain.user.dto.user.response.DuplicateNicknameResponse;
 import com.shy_polarbear.server.domain.user.exception.DuplicateNicknameException;
 import com.shy_polarbear.server.domain.auth.jwt.provider.KakaoProvider;
 import com.shy_polarbear.server.domain.user.service.UserService;
 import com.shy_polarbear.server.domain.auth.jwt.JwtDto;
-import com.shy_polarbear.server.domain.auth.repository.mysql.RefreshToken;
-import com.shy_polarbear.server.domain.auth.repository.mysql.RefreshTokenRepository;
 import com.shy_polarbear.server.global.auth.security.PrincipalDetails;
 import com.shy_polarbear.server.domain.auth.dto.request.JoinRequest;
 import com.shy_polarbear.server.domain.auth.dto.request.SocialLoginRequest;
